third_party/WebKit/LayoutTests/fast/dom/shadow/content-deep-expected.txt

::content /deep/ should match deep into distributed subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(target).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
